DSEGOG-67 moving channel options into a separate dialog
Still to customise the look of this dialog
